+++ tests/config/test_setup_store.py
