+++ tests/integrations/llm_providers/test_credential_kinds.py
